backend/requirements.txt
# --- Hashing password ---
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# --- JWT ---
PyJWT==2.8.0

# --- Email validation for Pydantic EmailStr ---
email-validator>=2.0.0

# --- FastAPI core ---
fastapi==0.110.0
starlette==0.36.3

# --- Pydantic v2 ---
pydantic==2.12.5
pydantic-core==2.41.5
annotated-types==0.6.0
pydantic-settings==2.12.0

# --- Server ---
uvicorn[standard]
click==8.1.7

# --- DB ---
sqlalchemy
psycopg
psycopg2-binary
pgvector

# --- Utils ---
httpx==0.26.0
python-dotenv
sentence-transformers
boto3==1.34.78
botocore==1.34.78
s3transfer==0.10.1
jmespath==1.0.1
six==1.16.0
python-dateutil==2.9.0.post0
alembic
prometheus-fastapi-instrumentator
numpy==1.26.4

# --- Testing ---
pytest
pytest-asyncio
pytest-mock
respx
